rename repository package to mongodb and pkg/mongodb to pkg/mongodbutils
