+++ app/api/infrastructure/mongodb/errors.go
@@ -1,4 +1,4 @@
-package repository
+package mongodb
 
 import (
 	"errors"
