+++ app/api/infrastructure/mongodb/user_activity.go
@@ -1,4 +1,4 @@
-package repository
+package mongodb
 
 import (
 	"context"
