+++ app/api/pkg/mongodbutils/constant.go
@@ -1,4 +1,4 @@
-package mongodb
+package mongodbutils
 
 // Err11000DupKeyError code error for E11000.
 const Err11000DupKeyError = 11000
